Move CallCounterProvider interface definition into resolver provider
