+++ xds/src/main/java/io/grpc/xds/XdsNameResolverProvider.java
@@ -24,6 +24,8 @@ import io.grpc.NameResolver.Args;
 import io.grpc.NameResolverProvider;
 import io.grpc.internal.ObjectPool;
 import java.net.URI;
+import java.util.concurrent.atomic.AtomicLong;
+import javax.annotation.Nullable;
 
 /**
  * A provider for {@link XdsNameResolver}.
@@ -75,5 +77,15 @@ public final class XdsNameResolverProvider extends NameResolverProvider {
 
   interface XdsClientPoolFactory {
     ObjectPool<XdsClient> getXdsClientPool() throws XdsInitializationException;
+  }
+
+  /**
+   * Provides the counter for aggregating outstanding requests per cluster:eds_service_name.
+   */
+  // TODO(chengyuanzhang): obtain a reference to the singleton instance in resolver and pass down
+  //  to LB policies in attributes.
+  // Introduced for testing.
+  interface CallCounterProvider {
+    AtomicLong getOrCreate(String cluster, @Nullable String edsServiceName);
   }
 }
